Add StreamingShufflePartialReadInvalidated event for streaming shuffle failure handling
- Added BlockManagerId import from org.apache.spark.storage package
- Implemented StreamingShufflePartialReadInvalidated case class after ShufflePushCompleted event
- Event contains shuffleId, mapId, and bmAddress to identify failed producer
- Enables partial read invalidation and upstream task recomputation for zero data loss guarantee
- Part of streaming shuffle feature implementation as per Agent Action Plan Section 0.6

+++ core/src/main/scala/org/apache/spark/scheduler/DAGSchedulerEvent.scala
@@ -145,6 +145,6 @@ private[scheduler] case class ShufflePushCompleted(
  */
 private[scheduler] case class StreamingShufflePartialReadInvalidated(
     shuffleId: Int,
-    mapId: Long,
+    mapId: Int,
     bmAddress: BlockManagerId)
   extends DAGSchedulerEvent
